fix: extract game helpers, fix isFinal missing "post" state
getWinner() in GameChip only checked state === "final", missing the
"post" state variant. This caused sports scores to not display
winner styling for games in "post" state. Also affected chipStatus()
and inline isFinal/isPre checks in the component body.

Extract isLive, isFinal, isPre, isCloseGame, getWinner, formatCountdown
to utils/gameHelpers.ts as the canonical source. Update GameChip,
SportsSummary, and ScrollrTicker to import from the shared module.
