+++ desktop/src/components/ScrollrTicker.tsx
@@ -5,7 +5,8 @@ import { useMotionValue, animate, AnimatePresence, motion } from "motion/react";
 import type { DashboardResponse, Trade, Game, RssItem, WidgetTickerData } from "../types";
 import type { MixMode, ChipColorMode, TickerDirection, ScrollMode, WidgetPinConfig } from "../preferences";
 import TradeChip from "./chips/TradeChip";
-import GameChip, { isLive, isCloseGame } from "./chips/GameChip";
+import GameChip from "./chips/GameChip";
+import { isLive, isCloseGame } from "../utils/gameHelpers";
 import RssChip from "./chips/RssChip";
 import FantasyChip from "./chips/FantasyChip";
 import ConsolidatedChip from "./chips/ConsolidatedChip";
