+++ desktop/src/components/dashboard/SportsSummary.tsx
@@ -11,7 +11,7 @@
  */
 import { useState, useMemo, useCallback } from "react";
 import { useScrollrCDC } from "../../hooks/useScrollrCDC";
-import { isCloseGame } from "../chips/GameChip";
+import { isLive, isFinal, isPre, isCloseGame, getWinner, formatCountdown } from "../../utils/gameHelpers";
 import { loadPref, savePref } from "../../preferences";
 import clsx from "clsx";
 import Tooltip from "../Tooltip";
@@ -32,26 +32,7 @@ function savePinned(pinned: PinnedMap): void {
 }
 
 // ── Game state helpers ──────────────────────────────────────────
-
-function isLive(g: Game): boolean {
-  return g.state === "in" || g.state === "in_progress";
-}
-
-function isFinal(g: Game): boolean {
-  return g.state === "final" || g.state === "post";
-}
-
-function isPre(g: Game): boolean {
-  return g.state === "pre";
-}
-
-function getWinner(g: Game): "home" | "away" | null {
-  if (!isFinal(g)) return null;
-  const a = Number(g.away_team_score);
-  const h = Number(g.home_team_score);
-  if (isNaN(a) || isNaN(h) || a === h) return null;
-  return h > a ? "home" : "away";
-}
+// Shared helpers imported from utils/gameHelpers.ts
 
 function gameStatus(game: Game): string {
   if (isLive(game)) return game.timer || game.status_short || "Live";
@@ -59,23 +40,6 @@ function gameStatus(game: Game): string {
   if (isPre(game)) return formatCountdown(game.start_time);
   if (game.state === "postponed") return "PPD";
   return "";
-}
-
-function formatCountdown(startTime: string): string {
-  const diff = new Date(startTime).getTime() - Date.now();
-  if (diff <= 0) return "Starting";
-  const h = Math.floor(diff / 3_600_000);
-  const m = Math.floor((diff % 3_600_000) / 60_000);
-  if (h >= 48) {
-    return new Date(startTime).toLocaleDateString(undefined, {
-      month: "short",
-      day: "numeric",
-    });
-  }
-  if (h >= 24) return "Tomorrow";
-  if (h > 0) return `in ${h}h ${m}m`;
-  if (m > 0) return `in ${m}m`;
-  return "Soon";
 }
 
 function abbreviate(name: string): string {
